ui added

+++ drug-store/src/main/java/kz/clouds/drugstore/ejb/basic/ControllerBase.java
@@ -1,5 +1,7 @@
 package kz.clouds.drugstore.ejb.basic;
 
+import java.util.List;
+
 import javax.annotation.Resource;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
@@ -22,8 +24,45 @@ public abstract class ControllerBase<T> {
 		}
 	}
 	
-	public void delete(T obj) {
-		em.remove(obj);
+	public void delete(Class<?> clazz, long id) {
+		try {
+			em = emf.createEntityManager();
+			Object obj = em.find(clazz, id);
+			if (obj!=null)
+				em.remove(obj);
+		}
+		finally {
+			if (em!=null)
+				em.close();
+		}
+	}
+	
+	public T get(Class<?> clazz, long id) {
+		try {
+			em = emf.createEntityManager();
+			Object obj = em.find(clazz, id);
+			if (obj!=null)
+				return (T) obj;
+			return null;
+		}
+		finally {
+			if (em!=null)
+				em.close();
+		}
+	}
+	
+	public List<T> getAll(Class<?> clazz, String order) {
+		try {
+			em = emf.createEntityManager();
+			List<Object> obj = em.createQuery("from "+clazz.getName()+" order by "+order).getResultList();
+			if (obj!=null)
+				return (List<T>) obj;
+			return null;
+		}
+		finally {
+			if (em!=null)
+				em.close();
+		}
 	}
 	
 }
